+++ src/main/java/com/github/hcsp/inheritance/Boy.java
@@ -1,20 +1,9 @@
 package com.github.hcsp.inheritance;
 
-public class Boy {
-    private String name;
-    private int age;
-
-    public Boy(String name, int age) {
-        this.name = name;
+public class Boy extends Human {
+    Boy(String name, int age) {
+        super(name);
         this.age = age;
-    }
-
-    public void sayMyName() {
-        System.out.println("我的名字是" + name);
-    }
-
-    public void sayMyAge() {
-        System.out.println("我的年龄是" + age);
     }
 
     public void sayBoy() {
